clutter/actor: Get next action from list before handling current action
Handling the current action might end up removing the current element
from the list, which would result in an invalid read when trying to get
the next element. Before glib 2.76 this was covered up by the use of the
slice allocator, but now this leads to a crash.

This was happening with the click action on the clear icon in the shell
search entry, which when clicked results in the clear icon being removed
along with the corresponding action.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/6552
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2955>