backends/native: Register X11 display with logind
logind has the ability to export the X11 display a session uses
with the logind session.

Previously, this mapping had to be set up at session opening time,
so we lost the feature when we moved to starting the X server
within the session instead of before the session.

Recently, logind gained a SetDisplay dbus method to allow the X11
display to be set after session open time.

This commit makes use of that new method.